Image requests now use https

+++ db/migrate/20201114013419_add_user_rating_to_movies.rb
@@ -1,5 +1,5 @@
 class AddUserRatingToMovies < ActiveRecord::Migration[5.2]
   def change
-    add_column :movies, :user_rating, :integer
+    add_column :movies, :user_rating, :decimal
   end
 end
